backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.27.0
